+++ src/cjs/draw.d.ts
@@ -173,13 +173,13 @@ export declare class drawutils implements DrawLib<void> {
     /**
      * Draw a rectangle.
      *
-     * @param {Vertex} position - The upper left corner of the rectangle.
+     * @param {XYCoords} position - The upper left corner of the rectangle.
      * @param {number} width - The width of the rectangle.
      * @param {number} height - The height of the rectangle.
      * @param {string} color - The color to use.
      * @param {number=1} lineWidth - (optional) The line with to use (default is 1).
      **/
-    rect(position: Vertex, width: number, height: number, color: string, lineWidth?: number): void;
+    rect(position: XYCoords, width: number, height: number, color: string, lineWidth?: number): void;
     _fillOrDraw(color: string): void;
     /**
      * Draw the given (cubic) bézier curve.
@@ -326,15 +326,15 @@ export declare class drawutils implements DrawLib<void> {
      * Note that if the x-scale and the y-scale are different the result will be a rectangle rather than a square.
      *
      * @method square
-     * @param {Vertex} center - The center of the square.
-     * @param {Vertex} size - The size of the square.
+     * @param {XYCoords} center - The center of the square.
+     * @param {number} size - The size of the square.
      * @param {string} color - The CSS color to draw the square with.
      * @param {number} lineWidth - The line with to use (optional, default is 1).
      * @return {void}
      * @instance
      * @memberof drawutils
      */
-    square(center: Vertex, size: number, color: string, lineWidth?: number): void;
+    square(center: XYCoords, size: number, color: string, lineWidth?: number): void;
     /**
      * Draw a grid of horizontal and vertical lines with the given (CSS-) color.
      *