+++ src/cjs/utils/dom/getAvailableContainerSpace.d.ts
@@ -1,0 +1,2 @@
+import { XYDimension } from "../../interfaces/core";
+export declare const getAvailableContainerSpace: (element: HTMLElement | SVGElement) => XYDimension;