+++ src/esm/drawgl.d.ts
@@ -260,7 +260,7 @@ export declare class drawutilsgl implements DrawLib<void> {
      * Note that if the x-scale and the y-scale are different the result will be a rectangle rather than a square.
      *
      * @method square
-     * @param {Vertex} center - The center of the square.
+     * @param {XYCords} center - The center of the square.
      * @param {Vertex} size - The size of the square.
      * @param {string} color - The CSS color to draw the square with.
      * @param {number=} lineWidth - (optional) The line width to use; default is 1.
@@ -268,7 +268,17 @@ export declare class drawutilsgl implements DrawLib<void> {
      * @instance
      * @memberof drawutils
      */
-    square(center: Vertex, size: number, color: string, lineWidth?: number): void;
+    square(center: XYCoords, size: number, color: string, lineWidth?: number): void;
+    /**
+     * Draw a rectangle.
+     *
+     * @param {XYCoords} position - The upper left corner of the rectangle.
+     * @param {number} width - The width of the rectangle.
+     * @param {number} height - The height of the rectangle.
+     * @param {string} color - The color to use.
+     * @param {number=1} lineWidth - (optional) The line with to use (default is 1).
+     **/
+    rect(position: XYCoords, width: number, height: number, color: string, lineWidth?: number): void;
     /**
      * Draw a grid of horizontal and vertical lines with the given (CSS-) color.
      *