+++ src/esm/utils/dom/getFProp.d.ts
@@ -1,0 +1,5 @@
+/**
+ * Internal helper function used to get 'float' properties from elements.
+ * Used to determine border withs and paddings that were defined using CSS.
+ */
+export declare const getFProp: (elem: HTMLElement | SVGElement, propName: string) => number;