+++ src/ts/PlotBoilerplate.ts
@@ -1460,6 +1460,7 @@ export class PlotBoilerplate {
    * Internal helper function used to get 'float' properties from elements.
    * Used to determine border withs and paddings that were defined using CSS.
    */
+  // TODO: this was moved to the DOM utils
   private getFProp(elem: HTMLElement | SVGElement, propName: string): number {
     return parseFloat(globalThis.getComputedStyle(elem, null).getPropertyValue(propName));
   }
@@ -1469,23 +1470,11 @@ export class PlotBoilerplate {
    *
    * Size minus padding minus border.
    **/
+  // TODO: this was moved to the DOM utils
   private getAvailableContainerSpace(): XYDimension {
     const _self: PlotBoilerplate = this;
     const container: HTMLElement = _self.canvas.parentNode as unknown as HTMLElement; // Element | Document | DocumentFragment;
-    // var canvas : HTMLCanvasElement = _self.canvas;
     _self.canvas.style.display = "none";
-    /* var
-	padding : number = parseFloat( globalThis.getComputedStyle(container, null).getPropertyValue('padding') ) || 0,
-	border : number = parseFloat( globalThis.getComputedStyle(_self.canvas, null).getPropertyValue('border-width') ) || 0,
-	pl : number = parseFloat( globalThis.getComputedStyle(container, null).getPropertyValue('padding-left') ) || padding,
-	pr : number = parseFloat( globalThis.getComputedStyle(container, null).getPropertyValue('padding-right') ) || padding,
-	pt : number = parseFloat( globalThis.getComputedStyle(container, null).getPropertyValue('padding-top') ) || padding,
-	pb : number = parseFloat( globalThis.getComputedStyle(container, null).getPropertyValue('padding-bottom') ) || padding,
-	bl : number = parseFloat( globalThis.getComputedStyle(_self.canvas, null).getPropertyValue('border-left-width') ) || border,
-	br : number = parseFloat( globalThis.getComputedStyle(_self.canvas, null).getPropertyValue('border-right-width') ) || border,
-	bt : number = parseFloat( globalThis.getComputedStyle(_self.canvas, null).getPropertyValue('border-top-width') ) || border,
-	bb : number = parseFloat( globalThis.getComputedStyle(_self.canvas, null).getPropertyValue('border-bottom-width') ) || border;
-	*/
     var padding: number = this.getFProp(container, "padding") || 0,
       border: number = this.getFProp(_self.canvas, "border-width") || 0,
       pl: number = this.getFProp(container, "padding-left") || padding,